requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.2
actionlib-msgs==4.9.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
ament-cmake-test==1.3.13
ament-copyright==0.12.15
ament-cppcheck==0.12.15
ament-cpplint==0.12.15
ament-flake8==0.12.15
ament-index-python==1.4.1
ament-lint==0.12.15
ament-lint-cmake==0.12.15
ament-package==0.14.2
ament-pep257==0.12.15
ament-uncrustify==0.12.15
ament-xmllint==0.12.15
angles==1.15.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
bond==4.1.2
builtin-interfaces==1.2.2
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
composition-interfaces==1.2.2
control-msgs==4.8.0
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
dataclasses-json==0.6.7
diagnostic-msgs==4.9.1
distro==1.9.0
domain-coordinator==0.10.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
fastapi==0.135.1
filelock==3.25.0
frozenlist==1.8.0
fsspec==2026.2.0
generate-parameter-library-py==0.6.0
geometry-msgs==4.9.1
greenlet==3.3.2
groq==1.0.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
idna==3.11
interactive-markers==2.3.2
Jinja2==3.1.6
joblib==1.5.3
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.16
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.9
laser-geometry==2.4.1
launch==1.0.14
launch-ros==0.19.13
launch-testing==1.0.14
launch-testing-ros==0.19.13
launch-xml==1.0.14
launch-yaml==1.0.14
lifecycle-msgs==1.2.2
map-msgs==2.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
message-filters==4.3.14
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nav-msgs==4.9.1
nav2-common==1.1.20
nav2-msgs==1.1.20
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openpyxl==3.1.5
orjson==3.11.7
ormsgpack==1.12.2
osrf-pycommon==2.1.6
packaging==26.0
pandas==2.3.3
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
pypdf==6.7.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-qt-binding==1.1.3
pytz==2025.2
PyYAML==6.0.3
rcl-interfaces==1.2.2
rclpy==3.3.20
rcutils==5.1.8
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
resource-retriever==3.1.3
rich==14.3.3
rmw-dds-common==1.6.0
ros2action==0.18.17
ros2bag==0.15.16
ros2cli==0.18.17
ros2component==0.18.17
ros2doctor==0.18.17
ros2interface==0.18.17
ros2launch==0.19.13
ros2lifecycle==0.18.17
ros2multicast==0.18.17
ros2node==0.18.17
ros2param==0.18.17
ros2pkg==0.18.17
ros2plugin==5.1.3
ros2run==0.18.17
ros2service==0.18.17
ros2topic==0.18.17
rosbag2-interfaces==0.15.16
rosbag2-py==0.15.16
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.8
rosidl-cli==3.1.8
rosidl-cmake==3.1.8
rosidl-generator-c==3.1.8
rosidl-generator-cpp==3.1.8
rosidl-generator-py==0.14.6
rosidl-parser==3.1.8
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.4
rosidl-typesupport-fastrtps-cpp==2.2.4
rosidl-typesupport-introspection-c==3.1.8
rosidl-typesupport-introspection-cpp==3.1.8
rpyutils==0.2.2
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sensor-msgs==4.9.1
sensor-msgs-py==4.9.1
sentence-transformers==5.2.3
shape-msgs==4.9.1
shellingham==1.5.4
six==1.17.0
slam-toolbox==2.6.10
smclib==4.1.2
sniffio==1.3.1
SQLAlchemy==2.0.47
sros2==0.10.9
starlette==0.52.1
statistics-msgs==1.2.2
std-msgs==4.9.1
std-srvs==4.9.1
stereo-msgs==4.9.1
sympy==1.14.0
tenacity==9.1.4
tf2-geometry-msgs==0.25.19
tf2-kdl==0.25.19
tf2-msgs==0.25.19
tf2-py==0.25.19
tf2-ros-py==0.25.19
tf2-sensor-msgs==0.25.19
tf2-tools==0.25.19
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
trajectory-msgs==4.9.1
transformers==5.2.0
triton==3.6.0
turtlesim==1.4.3
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unique-identifier-msgs==2.2.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
visualization-msgs==4.9.1
xacro==2.1.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
